requirements.txt
beautifulsoup4
requests
schedule
twilio
python-dotenv
playwright
sendgrid
